requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.1
fsspec==2025.7.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
onnxruntime==1.22.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.182
ultralytics-thop==2.0.16
urllib3==2.5.0
